% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcAtchison}
\alias{calcAtchison}
\title{Calculate Atchison distance of a matrix (among cols)}
\usage{
calcAtchison(B)
}
\arguments{
\item{B}{A matrix}
}
\value{
Atchison distance matrix
}
\description{
Calculate Atchison distance of a matrix (among cols)
}
