% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcdist}
\alias{calcdist}
\title{Calculate different types of distance (among cols)}
\usage{
calcdist(B, method = "pearson")
}
\arguments{
\item{B}{A matrix}

\item{method}{Method for distance matrix calculation, can be chosen from
"gower", "euclid", "pearson", "spearman" and "Atchison"}
}
\value{
Transformed matrix.
}
\description{
Calculate different types of distance (among cols)
}
