% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcGower}
\alias{calcGower}
\title{Calculate Gower distance (among cols)}
\usage{
calcGower(B)
}
\arguments{
\item{B}{A matrix for Gower distance calculation}
}
\value{
A distance matrix.
}
\description{
Calculate Gower distance (among cols)
}
