% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcProp}
\alias{calcProp}
\title{Calculate proportion of each component in gam model}
\usage{
calcProp(eNODAL)
}
\arguments{
\item{eNODAL}{An eNODAL object}
}
\value{
An eNODAL object with calculated proportion.
}
\description{
Calculate proportion of each component in gam model
}
