% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcProp_gam}
\alias{calcProp_gam}
\title{Calculate proportion of each component in gam model}
\usage{
calcProp_gam(eNODAL, index, Cl_sel)
}
\arguments{
\item{eNODAL}{An eNODAL object}

\item{index}{A number specify which column to calculate}

\item{Cl_sel}{A text specify which cluster this variable belongs to}
}
\value{
A vector contain proportion of variance each component explained.
}
\description{
Calculate proportion of each component in gam model
}
