% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calctrans}
\alias{calctrans}
\title{Transformation of distance matrix}
\usage{
calctrans(dists, method, d_max = 4)
}
\arguments{
\item{dists}{A matrix of distance}

\item{method}{transformation method, can be chosen from"pca" and "laplacian"
"pca" calculate the first several principal components and "laplacian" calculate last several eigen values.}

\item{d_max}{Maximum number of eigen decomposition. By default is 4.}
}
\value{
A matrix of co-occurrence.
}
\description{
Transformation of distance matrix
}
