% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_fac_num}
\alias{check_fac_num}
\title{Check each column of a dataframe is a factor or numerical}
\usage{
check_fac_num(X, max_level = 7)
}
\arguments{
\item{X}{A dataframe}

\item{max_level}{maximum number of levels,
i.e. if the unique value of a column less than max_level,
will set to be factor. By default is 7.}
}
\value{
a list contain checked matrix X and type of values for each output.
}
\description{
Check each column of a dataframe is a factor or numerical
}
