% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{clusterMethod}
\alias{clusterMethod}
\title{Run clustering method (do not need to specify number of clusters)}
\usage{
clusterMethod(Dist, clmethod, ...)
}
\arguments{
\item{Dist}{distance matrix to calculate}

\item{clmethod}{Method to run for clustering, can be chosen from:
apcluster: Affinity propagation clustering,
louvian: Louvian algorithm use knn graph,
dbscan: Density-based spatial clustering of applications with noise,
Dcutree: Adaptive branch pruning of hierarchical clustering.}

\item{...}{Other parameters can be passed to clusterMethod.}
}
\value{
A list contain clustering result.
}
\description{
Run clustering method (do not need to specify number of clusters)
}
