% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{createClParam}
\alias{createClParam}
\title{Create parameters list for stage II clustering.}
\usage{
createClParam(
  q0 = 0,
  knn.k = 40,
  dist_thresh = 0.4,
  eps = NULL,
  minPts = NULL,
  minClusterSize = NULL
)
}
\arguments{
\item{q0}{Parameter q0 in apcluster. Byt default is 0. If it is NULL, will not use apcluster.}

\item{knn.k}{Number of k-nearest neighbour to create knn graph in Louvian algorithm.
By default is 40, If it is NULL, will not use Louvian.}

\item{dist_thresh}{Threshold of distance when creating knn graph in Louvian algorithm.
By default is 0.4, If it is NULL, will not use Louvian.}

\item{eps}{Parameter eps in dbscan. If it is NULL, will not use dbscan.}

\item{minPts}{Parameter minPts in dbscan. If it is NULL, will not use dbscan.}

\item{minClusterSize}{Parameter minClusterSize in dbscan. If it is NULL, will not use dbscan.}
}
\value{
A list of methods and parameters.
}
\description{
Create parameters list for stage II clustering.
}
