% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{createConsensus}
\alias{createConsensus}
\title{Consensus clustering}
\usage{
createConsensus(Dist_list, Param0)
}
\arguments{
\item{Dist_list}{A list of distance matrix(see createDistList)}

\item{Param0}{Parameters for consensus clustering(see setParam)}
}
\value{
Consensus matrix.
}
\description{
Consensus clustering
}
