% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{createDist_base}
\alias{createDist_base}
\title{Run correlation distance}
\usage{
createDist_base(X_sel, dist_type)
}
\arguments{
\item{X_sel}{Matrix used to calculate correlation distance()}

\item{dist_type}{Type of correlation to calculate distance, can be chosen from
"pearson" and "spearman"}
}
\value{
Distance matrix
}
\description{
Run correlation distance
}
