% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{createDistList}
\alias{createDistList}
\title{Calculate list of distance matrix}
\usage{
createDistList(X_sel, V_exp = NULL, dist_method = "all")
}
\arguments{
\item{X_sel}{Matrix used to calculate correlation distance.}

\item{V_exp}{Matrix of explained variance (see calcProp).}

\item{dist_method}{Types of distance matrix to be calculated.
Can be chosen from "base" or "all".}
}
\value{
Weighted distance matrix
}
\description{
Calculate list of distance matrix
}
