% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{createParams}
\alias{createParams}
\title{Create parameters list for eNODAL object.}
\usage{
createParams(
  gam_k = 29,
  gam = TRUE,
  test_method = "F",
  h_adj = TRUE,
  adj_method = "BH",
  sig_level = c(LC = 0.01, Linear = 0.01, Sig = 0.01, Interaction = 0.05, Meta1 = 0.05,
    Meta2 = 0.05),
  adapt = FALSE,
  test_func = "gam",
  sig_test = "LC-test",
  save_dist = FALSE,
  q0 = 0,
  knn.k = 40,
  dist_thresh = 0.4,
  eps = NULL,
  minPts = NULL,
  minClusterSize = NULL,
  kcluster = NULL,
  consensus_param = list(clmethod = "louvian", knn.k = 15, dist_thresh = 0.75),
  ...
)
}
\arguments{
\item{gam_k}{Parameter k in GAM model. By default is 29.}

\item{gam}{Whether use create GAM formula. By default is TRUE.
If no continuous variable, set FALSE.}

\item{test_method}{Testing method.
Can be chosen from "F", "globaltest", "Tmax", "Chisq", "Cp". By default is "F"}

\item{h_adj}{Indicator of whether use hierarchical adjustment. By default is TRUE}

\item{adj_method}{Pvalue adjust method. See p.adjust. By default is "BH".}

\item{sig_level}{Significance level of different test. By default is set below:
LC = 0.01, Linear = 0.01, Interaction = 0.05, Meta1 = 0.05, Meta2 = 0.05.}

\item{adapt}{Indicator of whether to use adaptive way to decide testing type.}

\item{test_func}{Testing function used. By default is "gam".}

\item{sig_test}{Method for test between sig vs. non-sig. Can be choosen from "LC-test", "lm" or "gam.}

\item{save_dist}{Save distance result. By default is FALSE.}

\item{q0}{Parameter q0 in apcluster. Byt default is 0. If it is NULL, will not use apcluster.}

\item{knn.k}{Number of k-nearest neighbour to create knn graph in Louvian algorithm.
By default is 40, If it is NULL, will not use Louvian.}

\item{dist_thresh}{Threshold of distance when creating knn graph in Louvian algorithm.
By default is 0.4, If it is NULL, will not use Louvian.}

\item{eps}{Parameter eps in dbscan. If it is NULL, will not use dbscan.}

\item{minPts}{Parameter minPts in dbscan. If it is NULL, will not use dbscan.}

\item{minClusterSize}{Parameter minClusterSize in dbscan. If it is NULL, will not use dbscan.}

\item{kcluster}{Number of clusters if using fixed number of clustering method.}

\item{consensus_param}{Parameter list for consensus clustering.
Format: list(clmethod = ..., params = ...), available method and parameters listed above.
By default is Louvian, with knn.k = 15, dist_thresh = 0.3.}

\item{...}{Other parameters can be passed to eNODAL.}
}
\value{
A list contain parameters in eNODAL object.
}
\description{
Create parameters list for eNODAL object.
}
