% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{createwD2}
\alias{createwD2}
\title{Create distance matrix with respect to the variance explained.}
\usage{
createwD2(V_exp)
}
\arguments{
\item{V_exp}{Variance explained matrix (see calcProp).}
}
\value{
An eNODAL object with calculated Atchison distance.
}
\description{
Create distance matrix with respect to the variance explained.
}
