% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{createwDist}
\alias{createwDist}
\title{Calculate weighted distance}
\usage{
createwDist(X_sel, V_exp, Dist_base = NULL, D2 = NULL)
}
\arguments{
\item{X_sel}{Matrix used to calculate correlation distance.}

\item{V_exp}{Matrix of explained variance (see calcProp).}

\item{Dist_base}{Base distance, create from createDist_base.}

\item{D2}{Distance matrix for weight.}
}
\value{
Weighted distance matrix
}
\description{
Calculate weighted distance
}
