% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cvtCooccur}
\alias{cvtCooccur}
\title{Convert clustering result to co-occurrence matrix}
\usage{
cvtCooccur(cl_res)
}
\arguments{
\item{cl_res}{A vector of clustering membership}
}
\value{
A matrix of co-occurrence.
}
\description{
Convert clustering result to co-occurrence matrix
}
