% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.changename}
\alias{.changename}
\title{Change names of eNODAL clustering result}
\usage{
.changename(eNODAL)
}
\arguments{
\item{eNODAL}{An eNODAL object as input}
}
\value{
eNODAL object with names specified by Meta_name.
}
\description{
Change names of eNODAL clustering result
}
