% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.inhull}
\alias{.inhull}
\title{determine whether points is in convex hull}
\usage{
.inhull(
  testpts,
  calpts,
  hull = geometry::convhulln(calpts),
  tol = mean(mean(abs(calpts))) * sqrt(.Machine$double.eps)
)
}
\arguments{
\item{testpts}{A dataframe contain testing point}

\item{calpts}{A dataframe contain point to determine convex hull.}

\item{hull}{Create hull to test whether testpts in hull.
By default will create by geometry::convhulln(calpts).}

\item{tol}{Tolerance. By defualt is mean(mean(abs(calpts)))*sqrt(.Machine$double.eps)).}
}
\value{
An indicator weather each point belongs to the
convex hull spanned by calpts.
}
\description{
determine whether points is in convex hull
}
