% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{.nested_test}
\alias{.nested_test}
\title{Nested test.}
\usage{
.nested_test(data_g, calc0, calc1, func = lm, method = "F")
}
\arguments{
\item{data_g}{Matrix or data frame for model fitting.}

\item{calc0}{A string of formula for null hypothesis}

\item{calc1}{A string of formula for alternative hypothesis}

\item{func}{Fitting function used for test. Can be chosen from lm and gam.
By default is lm.}

\item{method}{Testing method.
Can be chosen from "F", "globaltest", "Tmax", "Chisq", "Cp". By default is "F"}
}
\value{
A list contain bootstrapping pvalue and fitted R.square
}
\description{
Nested test.
}
