% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{.npntest}
\alias{.npntest}
\title{nonparametric ANOVA test from Zhou's paper using bootstrap.}
\usage{
.npntest(data_g, calc0, calc1, func0 = lm, func1 = mgcv::gam, Boot = 50)
}
\arguments{
\item{data_g}{Matrix or data frame for model fitting.}

\item{calc0}{A string of formula for null hypothesis}

\item{calc1}{A string of formula for alternative hypothesis}

\item{func0}{Fitting function used for null hypothesis.}

\item{func1}{Fitting function used for alternative hypothesis.}

\item{Boot}{Number of bootstrapping. By default is 50.}
}
\value{
A list contain bootstrapping pvalue and fitted R.square
}
\description{
nonparametric ANOVA test from Zhou's paper using bootstrap.
}
