% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unify_class}
\alias{.unify_class}
\title{Unify all the type of X into matrix}
\usage{
.unify_class(X)
}
\arguments{
\item{X}{Input data.}
}
\value{
Matrix form of data.
}
\description{
Unify all the type of X into matrix
}
