% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{eNODAL_build}
\alias{eNODAL_build}
\title{Build eNODAL object}
\usage{
eNODAL_build(
  Z,
  Meta1,
  Meta2,
  Phenotype = NULL,
  Meta1_name = NULL,
  Meta2_name = NULL,
  ...
)
}
\arguments{
\item{Z}{A matrix/dataframe of omics data}

\item{Meta1}{A dataframe of first set of meta variables.}

\item{Meta2}{A dataframe of second set of meta variables.}

\item{Phenotype}{A dataframe of phenotype output.}

\item{Meta1_name}{Name of experiment conditions in Meta1. If it is NULL will name "Meta1".}

\item{Meta2_name}{Name of experiment conditions in Meta2. If it is NULL will name "Meta2".}

\item{...}{Other parameters can pass to eNODAL_build
See \code{\link{createParams}}}
}
\value{
An eNODAL object
}
\description{
Build eNODAL object
}
\examples{
data(Proteomics)
Z = Proteomics$Z
Meta1 = Proteomics$Meta1
Meta2 = Proteomics$Meta2
eNODAL_obj = eNODAL_build(Z, Meta1, Meta2)
}
