% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNODAL_example}
\alias{eNODAL_example}
\title{An example of fitted eNODAL object}
\format{
An object of class \code{eNODAL} of length 1.
}
\usage{
data(eNODAL_example, package = 'eNODAL')
}
\description{
An example of fitted eNODAL object
Generate from following code
data(Proteomics)
eNODAL_obj <- eNODAL_build(Z = Z, Meta1 = Meta1, Meta2 = Meta2,
sig_test = "lm", test_func = "lm")
eNODAL_obj <- runeNODAL(eNODAL_obj)
}
\keyword{datasets}
