% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fisher_trans}
\alias{fisher_trans}
\title{Fisher Transformation}
\usage{
fisher_trans(r)
}
\arguments{
\item{r}{correlation coefficient}
}
\value{
Fisher transformed coefficient
}
\description{
Fisher Transformation
}
