% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggSurface}
\alias{ggSurface}
\title{Function to make nice surfaces using ggplot}
\usage{
ggSurface(
  GAM,
  data,
  XYZ = NA,
  predict_val = NA,
  surf_min = NA,
  surf_max = NA,
  x.limits = NA,
  y.limits = NA,
  z.val = NA,
  exclude = NA,
  subtitle = "",
  lab_size = 3,
  nlevels = 5,
  contour_at = NA,
  skip = 0,
  axis_size = 15,
  axis_lab_size = 15
)
}
\arguments{
\item{GAM}{A gam function for prediction}

\item{data}{A dataframe contain all observed data}

\item{XYZ}{XY axis used for plot, Z axis will be used as median, can be multiple Z.}

\item{predict_val}{set additional predict parameters for gam fitting.
e.g. predict_val = data.frame(Meta2 = "control")}

\item{surf_min}{minimal value of the fitted surface.
If it is NA, the minimal value is determined by fitted value.}

\item{surf_max}{maximal value of the fitted surface.
If it is NA, the maximal value is determined by fitted value.}

\item{x.limits}{Limits of x axis.
If it is NA, the limits is determined by input data.}

\item{y.limits}{Limits of y axis.
If it is NA, the limits is determined by input data.}

\item{z.val}{Value of z to be plotted. Plot will condition on z.val and fit f(x,y,z=z.val).}

\item{exclude}{Point exclude from prediction used in predict.gam. By default is NA.}

\item{subtitle}{Subtitle of the figure.}

\item{lab_size}{label size in contour.}

\item{nlevels}{Number of levels in contour. By default is 5.}

\item{contour_at}{A vector indicates where to draw the contour.}

\item{skip}{Number of contours to skip.}

\item{axis_size}{axis.text size.}

\item{axis_lab_size}{axis.title size.}
}
\value{
A ggplot type object of NGF.
}
\description{
Function to make nice surfaces using ggplot
}
