% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_test.R
\name{lc_stat}
\alias{lc_stat}
\title{Calculate LC-Stats}
\usage{
lc_stat(D, M, Z, lmd = NULL)
}
\arguments{
\item{D}{Experimental matrix}

\item{M}{A matrix indicate which axis used in D to calculate LC}

\item{Z}{A vector of gene of interest.}

\item{lmd}{Scale factor to calculate weighted matrix in LC-Stats.}
}
\value{
A number of LC-Stats
}
\description{
Calculate LC-Stats
}
