% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_test.R
\name{LC_test0}
\alias{LC_test0}
\title{Calculate LC-Test}
\usage{
LC_test0(D, Z, n = 100)
}
\arguments{
\item{D}{Experimental matrix}

\item{Z}{A vector of gene of interest.}

\item{n}{Number of permutation}
}
\value{
Pvalue of LC-test
}
\description{
Calculate LC-Test
}
