% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{lm_test}
\alias{lm_test}
\title{Nested test for linear model.}
\usage{
lm_test(Y, Meta1, Meta2, calc0, calc1, method = "F")
}
\arguments{
\item{Y}{An omics feature as output.}

\item{Meta1}{A dataframe of first set of meta variables.}

\item{Meta2}{A dataframe of second set of meta variables.}

\item{calc0}{A string of formula for null hypothesis.}

\item{calc1}{A string of formula for alternative hypothesis.}

\item{method}{Testing method.
Can be chosen from "F", "globaltest", "Tmax", "Chisq", "Cp". By default is "F".}
}
\value{
A list contain bootstrapping pvalue and fitted R.square
}
\description{
Nested test for linear model.
}
