% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.knn.graph}
\alias{make.knn.graph}
\title{Make knn graph}
\usage{
make.knn.graph(Dist, knn.k = 15, dist_thresh = 0.3)
}
\arguments{
\item{Dist}{distance matrix}

\item{knn.k}{k-nearest neighbor. By default is 15}

\item{dist_thresh}{minimal distance threshold (k-nearest neighbor should smaller than dist_thresh)}
}
\value{
The corresponding knn graph
}
\description{
Make knn graph
}
