% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{NGFPlot}
\alias{NGFPlot}
\title{Generate NGF using eNODAL_obj object}
\usage{
NGFPlot(
  eNODAL_obj,
  Zaxis,
  XYaxis,
  By = NULL,
  scaled = T,
  type = NULL,
  formula0 = NULL,
  ...
)
}
\arguments{
\item{Zaxis}{An index or name of vector to select a variable for Zaxis.}

\item{XYaxis}{Variables of XYaxis,
can be chosen from all continuous variable in meta1 and meta2.
If more than one vector provided, will automatically do choose(n,2)}

\item{By}{A factor used for \code{by} in gam fitting.
If it is NULL, will not fit with by.}

\item{scaled}{Whether scale for Z variable.}

\item{type}{Can be chosen from "Int", "Add", "Meta1" and "Meta2",
for different types of visualization.
By default is NULL, will be chosen from eNODAL_obj fitting result.}

\item{formula0}{Formula used for gam fitting. By default is NULL.
Will automatically determined by other parameters.}

\item{...}{Other parameters can be passed to NGFPlot.}

\item{eNODAL}{An eNODAL_obj object}
}
\value{
A list of NGF plot.
}
\description{
Generate NGF using eNODAL_obj object
}
\examples{
data(eNODAL_example)
p = NGFPlot(eNODAL_example, 1, c("intake.P", "intake.C"))
}
\seealso{
\code{\link{ggSurface}}
}
