% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Proteomics}
\alias{Proteomics}
\title{Part of processed mouse liver nutriomics proteomics dataset}
\format{
A list of three type of data:
\describe{
\item{Z}{A dataframe of proteome data}
\item{Meta1}{A dataframe of nutrition intake condition in experiment}
\item{Meta2}{A factor of drug treatment}
}
}
\usage{
data(Proteomics, package = 'eNODAL')
}
\description{
A list of three type of data:
Proteomics data: A dataframe contain 127 mice and 1000 proteome.
Nutrition condition: A dataframe of 127 mice and 4 nutrition variables.
Drug treatment: A factor contain drug intake for each mouce.
}
\keyword{datasets}
