% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Proteomics_full}
\alias{Proteomics_full}
\title{Full proteomics mouse nutrition experimental data}
\format{
A list of three type of data:
\describe{
\item{Meta}{All phenotypical data}
\item{Prot_raw}{Raw proteomics data}
\item{Prot_RUV}{RUV processed proteomics data}
}
}
\usage{
data(Proteomics_full, package = 'eNODAL')
}
\description{
A list of three type of data:
}
\keyword{datasets}
