% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runAnno}
\alias{runAnno}
\title{Add annotation for variable name}
\usage{
runAnno(
  eNODAL,
  from = "UNIPROT",
  to = c("SYMBOL", "ENTREZID"),
  db = "org.Mm.eg.db"
)
}
\arguments{
\item{eNODAL}{An fitted eNODAL object.}

\item{from}{Initial id type in eNODAL object.}

\item{to}{To id type in eNODAL object.}

\item{db}{Annotation database.}
}
\value{
An eNODAL object.
}
\description{
Add annotation for variable name
}
