% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNODAL_run.R
\name{runeNODAL}
\alias{runeNODAL}
\title{Run eNODAL using with eNODAL object.}
\usage{
runeNODAL(eNODAL)
}
\arguments{
\item{eNODAL}{Input eNODAL object.}
}
\value{
eNODAL with clustering result.
}
\description{
Run eNODAL using with eNODAL object.
}
\examples{
data(Proteomics)
Z = Proteomics$Z
Meta1 = Proteomics$Meta1
Meta2 = Proteomics$Meta2
eNODAL_obj = eNODAL_build(Z, Meta1, Meta2, sig_test = "lm", test_func = "lm")
eNODAL_obj = runeNODAL(eNODAL_obj)
}
