% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage1.R
\name{runHT}
\alias{runHT}
\title{Run hypothesis testing using with eNODAL object.}
\usage{
runHT(eNODAL_obj, Boot = 50)
}
\arguments{
\item{eNODAL_obj}{Input eNODAL object.}

\item{Boot}{Number of Bootstrapping used in test (npntest). By default is 50.}
}
\value{
eNODAL object with clustering result(Stage I).
}
\description{
Run hypothesis testing using with eNODAL object.
}
