% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{runSubclust}
\alias{runSubclust}
\title{Run subclustering using with eNODAL_obj object.}
\usage{
runSubclust(eNODAL_obj)
}
\arguments{
\item{eNODAL_obj}{Input eNODAL object.}
}
\value{
eNODAL object with clustering result(Stage II).
}
\description{
Run subclustering using with eNODAL_obj object.
}
