% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_clusters}
\alias{show_clusters}
\title{Print all the clustering of eNODAL result}
\usage{
show_clusters(eNODAL_obj)
}
\arguments{
\item{eNODAL_obj, }{an eNODAL object.}
}
\value{
Clustering result of eNODAL object.
}
\description{
Print all the clustering of eNODAL result
}
