% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trans_clr}
\alias{trans_clr}
\title{Calculate centered log ratio(clr) transformation of a matrix}
\usage{
trans_clr(B)
}
\arguments{
\item{B}{A matrix for clr transformation}
}
\value{
Transformed matrix.
}
\description{
Calculate centered log ratio(clr) transformation of a matrix
}
