% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Boxplot}
\alias{Boxplot}
\title{Generate Boxplot using eNODAL_obj object}
\usage{
Boxplot(eNODAL_obj, XYaxis, Zaxis, scaled = F, ...)
}
\arguments{
\item{eNODAL_obj}{An eNODAL_obj object}

\item{XYaxis}{Variables of XYaxis,
can be chosen from all continuous variable in meta1 and meta2.
If more than one vector provided, will automatically do choose(n,2)}

\item{Zaxis}{An index or name of vector to select a variable for Zaxis.}

\item{scaled}{Whether scale for Z variable.}

\item{...}{Other parameters can be passed to NGFPlot.}
}
\value{
A list of NGF plot.
}
\description{
Generate Boxplot using eNODAL_obj object
}
\seealso{
\code{\link{ggSurface}}
}
