% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcProp}
\alias{calcProp}
\title{Calculate proportion of each component in gam model}
\usage{
calcProp(eNODAL_obj)
}
\arguments{
\item{eNODAL_obj}{An eNODAL_obj object}
}
\value{
An eNODAL_obj object with calculated proportion.
}
\description{
Calculate proportion of each component in gam model
}
