% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{Generate clustering result plot for eNODAL object}
\usage{
ClusterPlot(eNODAL)
}
\arguments{
\item{eNODAL}{An eNODAL object.(Already run two stage clustering)}
}
\value{
A concentric plot of eNODAL clustering result
}
\description{
Generate clustering result plot for eNODAL object
}
