% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.changename}
\alias{.changename}
\title{Change names of eNODAL_obj clustering result}
\usage{
.changename(eNODAL_obj)
}
\arguments{
\item{eNODAL_obj}{An eNODAL_obj object as input}
}
\value{
eNODAL_obj object with names specified by Meta_name.
}
\description{
Change names of eNODAL_obj clustering result
}
