% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_clusters}
\alias{get_clusters}
\title{Get clustering result of eNODAL object}
\usage{
get_clusters(eNODAL)
}
\arguments{
\item{eNODAL, }{an eNODAL object.}
}
\value{
Clustering result of eNODAL object.
}
\description{
Get clustering result of eNODAL object
}
