% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runAnno}
\alias{runAnno}
\title{Add annotation for variable name}
\usage{
runAnno(
  eNODAL_obj,
  from = "UNIPROT",
  to = c("SYMBOL", "ENTREZID"),
  db = "org.Mm.eg.db"
)
}
\arguments{
\item{eNODAL_obj}{An fitted eNODAL object.}

\item{from}{Initial id type in eNODAL object.}

\item{to}{To id type in eNODAL_obj object.}

\item{db}{Annotation database.}
}
\value{
An eNODAL_obj object.
}
\description{
Add annotation for variable name
}
