% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage2.R
\name{runSubclust}
\alias{runSubclust}
\title{Run subclustering using with eNODAL object.}
\usage{
runSubclust(eNODAL)
}
\arguments{
\item{eNODAL}{Input eNODAL object.}
}
\value{
eNODAL object with clustering result(Stage II).
}
\description{
Run subclustering using with eNODAL object.
}
